refactor(finance): migrate product price to BigDecimal

+++ src/main/java/com/bihariecart/service/impl/CartServiceImpl.java
@@ -89,7 +89,7 @@ public class CartServiceImpl implements CartService {
             newItem.setCart(cart);
             newItem.setProduct(product);
             newItem.setQuantity(requestedQuantity);
-            newItem.setPriceAtAddition(BigDecimal.valueOf(product.getPrice()));
+            newItem.setPriceAtAddition(product.getPrice());
             cart.addItem(newItem);
             cartItemRepository.save(newItem);
         }
